+++ wallet/tests/commands_for_test.go
@@ -757,6 +757,14 @@ func WalletRename(t *testing.T, args []string) error {
 	return err
 }
 
+func PassphraseUpdate(t *testing.T, args []string) error {
+	t.Helper()
+	argsWithCmd := []string{"passphrase", "update"}
+	argsWithCmd = append(argsWithCmd, args...)
+	_, err := ExecuteCmd(t, argsWithCmd)
+	return err
+}
+
 type ImportWalletResponse struct {
 	Wallet struct {
 		Name     string `json:"name"`